backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.18.1
pdfplumber>=0.11.0
plotly>=6.4.0
kaleido>=1.1.0
pandas>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.6.0
python-multipart>=0.0.9
